requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
pyyaml==6.0.2
apscheduler==3.10.4
python-dotenv==1.0.1
snscrape==0.7.0.20230622
